move indexHandler from plugin to algorithm
