+++ plugin/src/test/java/org/opensearch/ml/task/MLTaskManagerTests.java
@@ -39,7 +39,7 @@ import org.opensearch.core.index.shard.ShardId;
 import org.opensearch.ml.common.MLTask;
 import org.opensearch.ml.common.MLTaskState;
 import org.opensearch.ml.common.MLTaskType;
-import org.opensearch.ml.indices.MLIndicesHandler;
+import org.opensearch.ml.engine.indices.MLIndicesHandler;
 import org.opensearch.test.OpenSearchTestCase;
 import org.opensearch.threadpool.ThreadPool;
